Merge ViewDescriptors implementation (#6123)
## Summary

The motivation behind this PR was the lack of animation in the web
implementation when React.StrictMode was enabled. While debugging, I
discovered we had two similar mechanisms for storing connection between
components and animations: one for mobile animations and one for web
animations. I realized that the native implementation worked flawlessly
without any issues on the web as well, as it had been recently
refactored to align with React's guidelines. I decided to remove the
flawed web implementation and replace it with mobile implementation.

## Test plan

1. Open web-example and check if animations works.
2. Wrap any web example from example app with
`<React.StrictMode></React.StrictMode>` and see if animations works.