Remove unnecessary color processing in `interpolateColor` (#6213)
## Summary

Our `interpolateColor` function seems to have some unnecessary color
processing in it. It was added some 3-4 years ago and should not be
needed by now.

## Why

There is a function called `rgbaColor` which is used in `hsvToColor`
(which in turn is used only in `interpolateColor`) as well as directly
in `interpolateColor`. So it's safe to say it is used ***only*** there.
All it does is take r,g,b,a values, runs a small check on them and then
returns a rgba string.
It has a check whether we are on UI thread or not. If we were on UI, it
returned a number-formatted color (formatted for native platforms)
instead of the usual rgba string. The problem is, these values are then
directly returned from the function. I believe it was needed back then,
but now we have a function called `processColorInProps` which does
exactly that + it doesn't interfere with returned values of
`interpolateColor`. Taking all these things into consideration, I
believe this check in `rgbaColor` function is not needed anymore.

Moreover, it creates problems. When `interpolateColor` gets called on UI
thread (for example by using it inside `useDerivedValue` hook as shown
by this issue:
https://github.com/software-mansion/react-native-reanimated/issues/6119),
the user gets these number-formatted colors which don't work when passed
anywhere.

## Test plan

Copy `App.tsx` from snack linked to the mentioned issue and make sure
the text stays blue. Also, check out Color Interpolation example from
the example app.

Tested on iOS and Android (both Paper and Fabric) as well as Web. I've
noticed neither issues nor regressions.