Append `-Wpedantic` only for example apps (#6250)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR temporarily fixes
https://github.com/software-mansion/react-native-reanimated/issues/6234
by making the problematic `-Wpedantic` flag to be applied only in
Reanimated example apps so that we can monitor the soundness of the code
on the CI.

`-Wpedantic` flag was introduced in #6157 and released in 3.13.0.

## Test plan

Check if the CI is green.