Bring back useFrameCallback cleanup (#6323)
## Summary

https://github.com/software-mansion/react-native-reanimated/pull/6143
was supposed to improve cleanup of `useFrameCallback` but also deleted a
needed line - setting `callbackId` of callback registry to -1. Seems
like some code is dependent on this value and it caused some crashes
such as this one:
https://github.com/software-mansion/react-native-reanimated/pull/6143. I
haven't found a reliable way to reproduce the issue since we got the
info from issue creator's sentry. Nonetheless, analysis of the code has
shown that we need to set the `callbackId` to -1 on cleanup, otherwise
we perform operations that shouldn't be done anymore (e.g there's this
check:
https://github.com/software-mansion/react-native-reanimated/blob/6f6f73ae4550572ab6128b1c24c2195b9316edfc/packages/react-native-reanimated/src/frameCallback/FrameCallbackRegistryUI.ts#L105-L107).

## Test plan

As stated, the issue isn't really reproducible but the change doesn't
risk any regression and time will tell if it fixes the problem