Fix spotless config (#6349)
## Summary

We've noticed that sometimes Java files are not properly formatted but
the CI is still green. The reason is that only `src/main/java/**/*.java`
was formatted but there are still `paper/`, `fabric/` and
`reactNativeVersionPatch/` directories that are not covered by this
glob.

This PR extends the coverage of spotless formatter as well as formats
the remaining files.

## Test plan