Fix handling UTF-8 object keys (#6373)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

Regarding this
[issue](https://github.com/software-mansion/react-native-reanimated/issues/6220)
the PR solves it by creating a jsi key string from utf8.

Fixes #6220.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested on demo provided in issue mentioned above.

---------

Co-authored-by: Tomek Zawadzki <tomekzawadzki98@gmail.com>