fix: Chrome debugger crash caused by incorrect viewTag (#6437)
## Summary

The problem was reported in
[this](https://github.com/software-mansion/react-native-reanimated/issues/6432)
issue and
[this](https://github.com/software-mansion/react-native-reanimated/discussions/6359)
discussion seems to report the same problem.

In short, the crash was caused by incorrect check in `_getViewInfo` in
`createAnimatedComponent` which assigned the `_component` to the
`viewTag` only for web but not for chrome debugger. I replaced `IS_WEB`
check with `shouldBeUseWeb`, which also checks whether the current
environment is a chrome debugger.

## Test plan

- I created a bare app with `npx @react-native-community/cli init
ReactNativeDebugger --version 0.75.1`
- Installed latest reanimated and downloaded
[react-native-debugger](https://github.com/jhen0409/react-native-debugger?tab=readme-ov-file)
mentioned in
[this](https://github.com/software-mansion/react-native-reanimated/issues/6432)
issue
- Launched `react-native-debugger` and enabled remote debugging with js,
- When I tried to animate component with animated style, I saw the
reported crash