refactor: Less error prone `_value` in Mutables. (#6311)
## Summary

At the moment the `_value` property is an ordinary property, which makes
it visible via `Object.keys` or `const prop in sv`. This way sometimes
when it's accidentally passed to React it throws errors and we had to
add troubleshooting for that.

This PR makes it so it would be hidden via `enumerable: false`.

## Test plan

- [x] Runtime tests pass.