Skip zero delay (#6398)
## Summary

Closes
https://github.com/software-mansion/react-native-reanimated/issues/6341

Probably we want to fix the comparison inside `withDelay` to start the
delayed animation as soon as the conditions are fulfilled, instead of
waiting an additional frame.

```diff
-if (now - startTime > delayMs || animation.reduceMotion) {
+if (now - startTime >= delayMs || animation.reduceMotion) {
```

With this fix the delay of 0ms is only visible if you are really picky:

<table>
<tr>
<td> BEFORE </td>
<td> AFTER </td>
</tr>
<tr>
<td> 


https://github.com/user-attachments/assets/261d92d2-f302-4240-96c8-3d14352431c2

</td>
<td> 


https://github.com/user-attachments/assets/cf833359-7440-41d1-ab0a-daa3050bae15

</td>
</tr>

### Test plan

Includes simplification of tests of withDelay. Tested on both IOS and
Android